test: mark tests as integration
